Validators restructure
