+++ tests/integration/test_petstore.py
@@ -16,7 +16,8 @@ from openapi_core.schema.responses.models import Response
 from openapi_core.schema.schemas.models import Schema
 from openapi_core.schema.servers.models import Server, ServerVariable
 from openapi_core.shortcuts import create_spec
-from openapi_core.validators import RequestValidator, ResponseValidator
+from openapi_core.validation.request.validators import RequestValidator
+from openapi_core.validation.response.validators import ResponseValidator
 from openapi_core.wrappers import MockRequest, MockResponse
